Remove FloatCF, rename ComplexCF to CorrelationFunction, add logic to cast output as real if no outputs were complex
